requirements.txt
beautifulsoup4==4.9.3
gql==3.0.0a5
lxml==4.6.2
pymongo==3.11.2
python-dateutil==2.8.1
requests==2.25.1
